Gallery fixes
I rewrote the layout code to be simpler. Now it works. I have no idea why the old code wasn't working. I even tried copying and pasting the example. I am using `ui_columns` now and it works. It's not perfect but it's good enough for now.

It has some jank, which shouldn't be happening. Perhaps something is blocking and that should never ever ever happen. I will have to look into `egui_infinite_scroll` and the loader to see what's going on. I am using the `egui::Image` directly without the `thumbhash` the example uses. This means a spinner is shown instead of a thumbhash but that shouldn't block.